# server-spring-boot/api-specifications.yaml
openapi: 3.0.1
info:
  title: Hostel Management System API (MongoDB)
  description: RESTful API for managing hostel operations including student registration, room allocation, food menu management, and room change requests. This version uses MongoDB as the database.
  contact:
    name: Hostel Management Team
    url: https://hostel.com
    email: admin@hostel.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Development Server (MongoDB)
- url: https://api.hostel.com
  description: Production Server (MongoDB)
security:
- Bearer Authentication: []
paths:
...
  /api/student/room-change-request:
    post:
      tags:
      - Student Operations
      summary: Submit Room Change Request
      description: Submit a request to change current room assignment with reason
      operationId: submitRoomChangeRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomChangeRequestDto'
        required: true
      responses:
        201:
          description: Room change request submitted successfully
          content:
            application/json:
              schema:
                example: {"message": "Room change request submitted successfully"}
        400:
          description: Invalid request data
          content:
            application/json:
              schema:
                example: {"error": "Invalid room number"}
        401:
          description: Unauthorized - Student access required
          content:
            application/json:
              schema:
                example: {"error": "Unauthorized"}
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                example: {"error": "Database error"}
      security:
      - Bearer Authentication: []
  /api/rooms:
    get:
      tags:
      - Room Management
      summary: Get All Rooms
      description: Retrieve all rooms with their occupancy statistics and availability status
      operationId: getAllRooms
      responses:
        200:
          description: Rooms retrieved successfully
          content:
            application/json:
              schema:
                example: [{"roomNumber": "R001", "capacity": 4, "occupied": 2, "available": 2}]
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                example: {"error": "Database error"}
  /api/food-menu:
    get:
      tags:
      - Food Menu
      summary: Get Food Menu
      description: Retrieve the weekly food menu with meals organized by day and meal type
      operationId: getFoodMenu
      responses:
        200:
          description: Food menu retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FoodMenu'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                example: {"error": "Database error"}
components:
  schemas:
...
    RoomChangeRequestDto:
      required:
      - newRoomNumber
      - reason
      type: object
      properties:
        newRoomNumber:
          type: string
          description: Desired new room number
        reason:
          type: string
          description: Reason for room change request
    FoodMenu:
      type: object
      properties:
        id:
          type: string
          description: Menu item unique identifier
        day:
          type: string
          description: Day of the week
        mealType:
          type: string
          description: Type of meal (breakfast/lunch/dinner)
        items:
          type: array
          items:
            type: string
          description: List of food items for the meal
        nutritionalInfo:
          type: string
          description: Nutritional information
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Enter JWT Bearer token in the format - Bearer {token}
      scheme: bearer
      bearerFormat: JWT 